First pass at cookbook-openstack-integration-test
This first pass gets tempest installed to /opt/tempest, and sets up
necessary users for testing.  More work needs to be done on adding
correct configuration to tempest.conf and enabling different features
depending on what exists in the environment.  This will most likely
involve exposing everything via different attributes.

Since we checkout master of tempest, the cookbook assumes tempest
will be executed to use a virtual env.  This will hopefully safe-guard
us from future package-related issues.  Note that at the moment
the python virtual env does NOT build cleanly on RHEL 6.5.  According
to devstack, it appears that a number of work-arounds are required to
make tempest run on RHEL 6.5.

Change-Id: I08224d2f4784d2fc041a5806f221e7411d5b813a