Remove bind cookbook from Berksfile and fix version pin
The bind cookbook has been migrated to the Sous Chefs and a new release has
finally been done which removes the need to have this in the pointer (which is
now broken).

In addition, update cacert which is shipped in the python requests package on
RHEL to work around expired LetsEncrypt root certificate.

Depends-On: https://review.opendev.org/813719
Change-Id: Id537867e47ed54c7028abf7037b251077e57de64
Signed-off-by: Lance Albertson <lance@osuosl.org>